# Always mock SubscriptionRegistry & DiscoveryResponder for wemo tests