--- a/tests/components/wemo/conftest.py
+++ b/tests/components/wemo/conftest.py
@@ -22,8 +22,8 @@ def pywemo_model_fixture():
     return "LightSwitch"
 
 
-@pytest.fixture(name="pywemo_registry")
-def pywemo_registry_fixture():
+@pytest.fixture(name="pywemo_registry", autouse=True)
+async def async_pywemo_registry_fixture():
     """Fixture for SubscriptionRegistry instances."""
     registry = create_autospec(pywemo.SubscriptionRegistry, instance=True)
 
@@ -38,6 +38,14 @@ def pywemo_registry_fixture():
 
     with patch("pywemo.SubscriptionRegistry", return_value=registry):
         yield registry
+
+
+@pytest.fixture(name="pywemo_discovery_responder", autouse=True)
+def pywemo_discovery_responder_fixture():
+    """Fixture for the DiscoveryResponder instance."""
+    discovery_responder = create_autospec(pywemo.ssdp.DiscoveryResponder, instance=True)
+    with patch("pywemo.ssdp.DiscoveryResponder", return_value=discovery_responder):
+        yield discovery_responder
 
 
 @pytest.fixture(name="pywemo_device")